requirements.txt
anyio==3.6.1
blinker==1.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.6.2
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
orjson==3.8.0
pydantic==1.10.2
pymongo==3.11.3
python-dotenv==0.16.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
Werkzeug==1.0.1
